Move MockScene to utils.py
